refactor: change file name to the correct ones
