+++ src/pages/Upload/UploadInProgressStep/index.tsx
@@ -17,7 +17,7 @@ import BottomButtonBar from '../../../components/common/BottomButtonBar';
 import { PageContainer } from '../../Home/index.styles';
 import sendFile from '../../../apis/assignment';
 import useAuthStore from '../../../store/useAuthStore';
-import useMaterialStore from '../../../store/materialStore';
+import useMaterialStore from '../../../store/useMaterialStore';
 import useModal from '../../../hooks/common/useModal';
 import Modal from '../../../components/common/Modal';
 import { useNextStep } from '..';
